+++ docs/src/release_notes.rst
@@ -17,6 +17,7 @@ Please delete the cache file before building documentation.
 - Enable configurations without autodoc (:issue:`48`)
 - Fix crash on annotation-only assignment (:issue:`50`)
 - Fix issue with filenames that have dots (:issue:`52`)
+- Correctly remove extension when building non-HTML documentation (:issue:`53`)
 
 0.4.0 (2021-10-08)
 ------------------